fix(skills): dwsplashrange is bigint — read as i64, not i32
The previous fix exposed the real bug: dwsplashrange is bigint in PG but
the mapper read i32 (tokio-postgres refuses bigint->i32). load_all failed
at boot (swallowed: 'skills desactivadas') and load(skill_id) panicked the
channel right after a successful world entry — killing :30003 every time
the player used a skill.

Now read as i64 (max(0) -> u32). Verified vs live PG: 33 skills have real
splash (range up to 600). The only remaining i32 reads are vnum and
dwtargetrange (both integer — correct).

Workspace 712 passed / 0 failed.

+++ source/reforge/game_core/src/skill.rs
@@ -575,7 +575,7 @@ fn skill_proto_from_row(r: &tokio_postgres::Row) -> Option<SkillProto> {
         attr_type: attr_type_from_text(r.get::<_, Option<String>>(11).as_deref().unwrap_or("")),
         max_hit: r.get::<_, i16>(12) as u16,
         target_range: r.get::<_, i32>(13) as u32,
-        splash_range: r.get::<_, i32>(14).max(0) as u32,
+        splash_range: r.get::<_, i64>(14).max(0) as u32,
         splash_adjust_poly: r.get::<_, Option<String>>(15).unwrap_or_default(),
     })
 }
